// types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      captions: {
        Row: {
          caption_request_id: number | null
          content: string | null
          created_datetime_utc: string
          humor_flavor_id: number | null
          id: string
          image_id: string
          is_featured: boolean
          is_public: boolean
          like_count: number
          llm_prompt_chain_id: number | null
          modified_datetime_utc: string | null
          profile_id: string
        }
        Insert: {
          caption_request_id?: number | null
          content?: string | null
          created_datetime_utc?: string
          humor_flavor_id?: number | null
          id?: string
          image_id: string
          is_featured?: boolean
          is_public: boolean
          like_count?: number
          llm_prompt_chain_id?: number | null
          modified_datetime_utc?: string | null
          profile_id: string
        }
        Update: {
          caption_request_id?: number | null
          content?: string | null
          created_datetime_utc?: string
          humor_flavor_id?: number | null
          id?: string
          image_id?: string
          is_featured?: boolean
          is_public?: boolean
          like_count?: number
          llm_prompt_chain_id?: number | null
          modified_datetime_utc?: string | null
          profile_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "captions_caption_request_id_fkey"
            columns: ["caption_request_id"]
            isOneToOne: false
            referencedRelation: "caption_requests"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "captions_humor_flavor_id_fkey"
            columns: ["humor_flavor_id"]
            isOneToOne: false
            referencedRelation: "humor_flavors"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "captions_image_id_fkey"
            columns: ["image_id"]
            isOneToOne: false
            referencedRelation: "images"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "captions_llm_prompt_chain_id_fkey"
            columns: ["llm_prompt_chain_id"]
            isOneToOne: false
            referencedRelation: "llm_prompt_chains"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "captions_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      profiles: {
        Row: {
          created_datetime_utc: string | null
          email: string | null
          first_name: string | null
          id: string
          is_in_study: boolean
          is_matrix_admin: boolean
          is_superadmin: boolean
          last_name: string | null
          modified_datetime_utc: string | null
        }
        Insert: {
          created_datetime_utc?: string | null
          email?: string | null
          first_name?: string | null
          id: string
          is_in_study?: boolean
          is_matrix_admin?: boolean
          is_superadmin?: boolean
          last_name?: string | null
          modified_datetime_utc?: string | null
        }
        Update: {
          created_datetime_utc?: string | null
          email?: string | null
          first_name?: string | null
          id?: string
          is_in_study?: boolean
          is_matrix_admin?: boolean
          is_superadmin?: boolean
          last_name?: string | null
          modified_datetime_utc?: string | null
        }
        Relationships: []
      }
// ...
      shares: {
        Row: {
          caption_id: string | null
          created_datetime_utc: string
          id: number
          profile_id: string | null
          proper_destination: string | null
          share_to_destination_id: number | null
        }
        Insert: {
          caption_id?: string | null
          created_datetime_utc?: string
          id?: number
          profile_id?: string | null
          proper_destination?: string | null
          share_to_destination_id?: number | null
        }
        Update: {
          caption_id?: string | null
          created_datetime_utc?: string
          id?: number
          profile_id?: string | null
          proper_destination?: string | null
          share_to_destination_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "shares_caption_id_fkey"
            columns: ["caption_id"]
            isOneToOne: false
            referencedRelation: "captions"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "shares_caption_id_fkey"
            columns: ["caption_id"]
            isOneToOne: false
            referencedRelation: "v_richest_image_dedup"
            referencedColumns: ["caption_id"]
          },
          {
            foreignKeyName: "shares_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "shares_share_to_destination_id_fkey"
            columns: ["share_to_destination_id"]
            isOneToOne: false
            referencedRelation: "share_to_destinations"
            referencedColumns: ["id"]
          },
        ]
      }
      sidechat_posts: {
        Row: {
          content: string | null
          created_datetime_utc: string
          id: string
          like_count: number
          post_datetime_utc: string
        }
        Insert: {
          content?: string | null
          created_datetime_utc?: string
          id: string
          like_count?: number
          post_datetime_utc: string
        }
        Update: {
          content?: string | null
          created_datetime_utc?: string
          id?: string
          like_count?: number
          post_datetime_utc?: string
        }
        Relationships: []
      }
// ...
      terms: {
        Row: {
          created_datetime_utc: string
          definition: string
          example: string
          id: number
          modified_datetime_utc: string | null
          priority: number
          term: string
          term_type_id: number | null
        }
        Insert: {
          created_datetime_utc?: string
          definition: string
          example: string
          id?: number
          modified_datetime_utc?: string | null
          priority?: number
          term: string
          term_type_id?: number | null
        }
        Update: {
          created_datetime_utc?: string
          definition?: string
          example?: string
          id?: number
          modified_datetime_utc?: string | null
          priority?: number
          term?: string
          term_type_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "terms_term_type_id_fkey"
            columns: ["term_type_id"]
            isOneToOne: false
            referencedRelation: "term_types"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}